refactor(templates): split generated component helpers
Extract Chat styles and API Playground helpers into registered generated modules. Keep static assertions aligned with the new imports and registration paths.
