+++ src/lib/structures.ts
@@ -16,6 +16,7 @@ import { notFoundTemplate } from "../templates/app/not-found.js";
 import { themeTemplate } from "../templates/app/theme.js";
 
 import { chatTemplate } from "../templates/components/Chat.js";
+import { chatStylesTemplate } from "../templates/components/ChatStyles.js";
 import { lockBodyScrollTemplate } from "../templates/components/LockBodyScroll.js";
 import { docsTemplate } from "../templates/components/Docs.js";
 import { docsSideBarTemplate } from "../templates/components/DocsSideBar.js";
@@ -100,6 +101,7 @@ import { configTemplate } from "../templates/utils/config.js";
 import { playgroundAllowlistTemplate } from "../templates/utils/playgroundAllowlist.js";
 import { ssrfGuardTemplate } from "../templates/utils/ssrfGuard.js";
 import { apiSnippetsTemplate } from "../templates/utils/apiSnippets.js";
+import { apiPlaygroundUtilsTemplate } from "../templates/utils/apiPlayground.js";
 import { requestBodyTemplate } from "../templates/utils/requestBody.js";
 import { playgroundRoutesTemplate } from "../templates/app/api/playground/route.js";
 import { playgroundAllowlistStubTemplate } from "../templates/services/openapi/playgroundAllowlistStub.js";
@@ -223,9 +225,11 @@ export const appStructure: Record<string, string> = {
   "utils/playgroundAllowlist.ts": playgroundAllowlistTemplate,
   "utils/ssrfGuard.ts": ssrfGuardTemplate,
   "utils/apiSnippets.ts": apiSnippetsTemplate,
+  "utils/apiPlayground.ts": apiPlaygroundUtilsTemplate,
   "utils/requestBody.ts": requestBodyTemplate,
 
   "components/Chat.tsx": chatTemplate,
+  "components/ChatStyles.ts": chatStylesTemplate,
   "components/LockBodyScroll.ts": lockBodyScrollTemplate,
   "components/Docs.tsx": docsTemplate,
   "components/DocsSideBar.tsx": docsSideBarTemplate,
